refactor: 토너먼트 및 매치 파일 디렉토리로 분리
